thay doi map, cap nhat bar và sound

+++ src/uet/oop/bomberman/entities/Entity.java
@@ -20,8 +20,8 @@ public abstract class Entity {
     protected int layer;
     int time = 0;
     protected boolean alive;
-    public int centerX = x + Sprite.SCALED_SIZE / 2;
-    public int centerY = y + Sprite.SCALED_SIZE / 2;
+    public float centerX = x + Sprite.SCALED_SIZE / 2;
+    public float centerY = y + Sprite.SCALED_SIZE / 2;
 
     public boolean isAlive() {
         return alive;
